+++ CHANGELOG.md
@@ -6,8 +6,10 @@
   json serializable of models and added copyWith method (Message, Reaction and Reply Message).
 * **Breaking**: [181](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/181) Removed
   deprecated field `showTypingIndicator` from ChatView.
-* **Breaking**: [187](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/187) update sendBy
-    parameter name of Message class to sentBy
+* **Breaking**: [188](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/188) update
+  sendBy parameter name of Message class to sentBy
+* **Breaking**: [190](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/190) Move
+  currentUser into ChatController from ChatView widget and rename chatUsers to otherUsers
 * **Feat**: [179](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/179) Added reply
   suggestions functionality
 * **Feat**: [157](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/157) Added onTap